+++ d3/colors.js
@@ -8,13 +8,13 @@ var colorFederated = function(treeData) {
         if (d.tag && d.tag === 'fed-op') {
             if (d.properties && d.properties.connection) {
                 d.federated = d.properties.connection.split(".")[0];
+                d.nodeClass = d.properties.connection.split(".")[0];
             }
         } else if (d.parent && d.parent.federated) {
             d.federated = d.parent.federated;
+            d.nodeClass = d.parent.federated;
         }
-    }, function(d) {
-        return d.children && d.children.length > 0 ? d.children : null;
-    });
+    }, common.allChildren);
 };
 
 exports.colorFederated = colorFederated;
